Close #14
I added to the README.md file the technique we are using to determine
the Forward/Inverse Velocity of the Delta Robot. This is currently
derived in MatlabScript.md and will be ported over to C code in the near
future. An issue will be opened to port this technique over.